Remove Automatic Scaling strategy option
