--- frontend/public/extend/devconsole/components/import/Scaling.scss
@@ -1,7 +1,0 @@
-.warning-text {
-  color: var(--pf-global--warning-color--100);
-}
-
-.odc-replicas {
-  max-width: 800px !important;
-}
