+++ frontend/public/extend/devconsole/shared/components/advanced-options/Scaling.scss
@@ -1,0 +1,8 @@
+.odc-scaling {
+  font-size: 18px;
+}
+
+.odc-scaling-info {
+  font-size: 15px;
+  font-weight: bolder;
+}
